Generic, recursive input selection
