+++ src/Wallet/errors/index.ts
@@ -1,1 +1,2 @@
 export { InsufficientValueInUtxosForSelection } from './InsufficientValueInUtxosForSelection'
+export { MaximumTransactionInputsExceeded } from './MaximumTransactionInputsExceeded'
