+++ src/Wallet/lib/index.ts
@@ -1,4 +1,4 @@
-export * from './coin_selection'
+export * from './select_inputs'
 export * from './address_derivation'
 export * from './address_discovery'
 export * from './get_next_address'
