+++ src/lib/bindings/index.ts
@@ -13,3 +13,19 @@ export function getBindingsForEnvironment () {
     ? cardanoNodeBindings
     : require('cardano-wallet-browser') as typeof cardanoNodeBindings
 }
+
+export function getRandomBytesForEnvironmentAsHex (byteLength = 32): string {
+  if (isNode()) {
+    const crypto = require('crypto')
+    const buf = crypto.randomBytes(byteLength)
+    return buf.toString('hex')
+  } else {
+    const byteArray = new Uint8Array(byteLength)
+    const w = window as any
+    const randomBytes = w.crypto.getRandomValues(byteArray)
+    return Array
+      .from(randomBytes)
+      .map((b: any) => b.toString(16).padStart(2, '0'))
+      .join('')
+  }
+}
